+++ app/Http/Controllers/Dosen/SuratIzinController.php
@@ -9,7 +9,7 @@ use Auth;
 
 class SuratIzinController extends Controller
 {
-    public function read(Request $request)
+    public function read(Request $request, $id_prodi = null)
     {
          $entries = $request->input('entries', 5);
         $user = Auth::user();
@@ -17,12 +17,14 @@ class SuratIzinController extends Controller
         // Ambil ID dosen login
         $id_dosen = $user->id_dosen ?? $user->id;
 
-        $suratizin = DB::table('surat_izin')
+        $query = DB::table('surat_izin')
             ->join('mahasiswa', 'surat_izin.id_mahasiswa', '=', 'mahasiswa.id')
+            ->join('prodi', 'surat_izin.id_prodi', '=', 'prodi.id')
             ->leftJoin('dosen as dosen1', 'surat_izin.dosen_pembimbing_1', '=', 'dosen1.id')
             ->leftJoin('dosen as dosen2', 'surat_izin.dosen_pembimbing_2', '=', 'dosen2.id')
             ->select(
                 'surat_izin.*',
+                'prodi.nama as prodi',
                 'mahasiswa.nama as nama',
                 'mahasiswa.nobp',
                 'dosen1.nama as dosen_pembimbing_1',
@@ -32,15 +34,25 @@ class SuratIzinController extends Controller
                 $query->where('surat_izin.dosen_pembimbing_1', $id_dosen)
                       ->orWhere('surat_izin.dosen_pembimbing_2', $id_dosen);
             })
-            ->orderBy('surat_izin.id', 'DESC')
-            ->paginate($entries);
+            ->orderBy('surat_izin.id', 'DESC');
 
+         // Filter berdasarkan prodi
+        if ($id_prodi) {
+            $query->where('surat_izin.id_prodi', $id_prodi);
+        }
+
+
+        // Supaya pagination tetap bawa query string (search / entries)
+        $suratizin = $query->paginate($entries);
         $suratizin->appends($request->all());
 
-        return view('dosen.suratizin.index', compact('suratizin'));
+        $username = auth()->user()->username;
+        $dosen = DB::table('dosen')->where('nip', $username)->first();
+
+        return view('dosen.suratizin.index', compact('suratizin','dosen','id_prodi'));
     }
 
-    public function feature(Request $request)
+    public function feature(Request $request, $id_prodi = null)
     {
         // Ambil data user login
         $user = Auth::user();
@@ -51,10 +63,12 @@ class SuratIzinController extends Controller
         // Base query
         $query = DB::table('surat_izin')
             ->join('mahasiswa', 'surat_izin.id_mahasiswa', '=', 'mahasiswa.id')
+            ->join('prodi', 'surat_izin.id_prodi', '=', 'prodi.id')
             ->leftJoin('dosen as dosen1', 'surat_izin.dosen_pembimbing_1', '=', 'dosen1.id')
             ->leftJoin('dosen as dosen2', 'surat_izin.dosen_pembimbing_2', '=', 'dosen2.id')
             ->select(
                 'surat_izin.*',
+                'prodi.nama as prodi',
                 'mahasiswa.nama as nama',
                 'mahasiswa.nobp',
                 'dosen1.nama as dosen_pembimbing_1',
@@ -64,6 +78,11 @@ class SuratIzinController extends Controller
                 $q->where('surat_izin.dosen_pembimbing_1', $id_dosen)
                   ->orWhere('surat_izin.dosen_pembimbing_2', $id_dosen);
             });
+
+         // Filter berdasarkan prodi
+        if ($id_prodi) {
+            $query->where('surat_izin.id_prodi', $id_prodi);
+        }
 
         // 🔍 Fitur search
         if ($request->filled('search')) {
@@ -91,14 +110,17 @@ class SuratIzinController extends Controller
         // Bawa query string ke pagination
         $suratizin->appends($request->all());
 
+        $username = auth()->user()->username;
+        $dosen = DB::table('dosen')->where('nip', $username)->first();
+
         // Tampilkan ke view
-        return view('dosen.suratizin.index', compact('suratizin'));
+        return view('dosen.suratizin.index', compact('suratizin','dosen','id_prodi'));
     } 
 
-    public function add(){
+    public function add($id_prodi = null){
         $mahasiswa = DB::table('mahasiswa')->orderBy('id','DESC')->get();
         $dosen = DB::table('dosen')->orderBy('id','DESC')->get();
-        return view('dosen.suratizin.create',['mahasiswa'=>$mahasiswa,'dosen'=>$dosen]);
+        return view('dosen.suratizin.create',['mahasiswa'=>$mahasiswa,'dosen'=>$dosen,'id_prodi'=>$id_prodi]);
     }
 
     public function create(Request $request){
@@ -109,17 +131,17 @@ class SuratIzinController extends Controller
             'judul_penelitian' => 'required|string|max:255',
             'dosen_pembimbing_1' => 'required|exists:dosen,nama',
             'dosen_pembimbing_2' => 'required|exists:dosen,nama',
-            'isi_surat' => 'required|string',
+            'isi_surat' => 'nullable|string',
         ],[
             'jenis.required' => 'Jenis wajib diisi.',
             'id_mahasiswa.exists' => 'Mahasiswa yang dipilih tidak valid..',
             'judul_penelitian.required' => 'Judul Penelitian wajib diisi.',
             'dosen_pembimbing_1.exists' => 'Dosen Bimbingan 1 yang dipilih tidak valid..',
             'dosen_pembimbing_2.exists' => 'Dosen Bimbingan 2 yang dipilih tidak valid..',
-            'isi_surat.required' => 'Isi surat wajib diisi.',
         ]);
         DB::table('surat_izin')->insert([  
             'jenis' => $request->jenis,
+            'id_prodi' => $request->id_prodi,
             'id_mahasiswa' => $request->id_mahasiswa,
             'judul_penelitian' => $request->judul_penelitian,
             'dosen_pembimbing_1' => $request->dosen_pembimbing_1,
@@ -127,7 +149,7 @@ class SuratIzinController extends Controller
             'isi_surat' => $request->isi_surat
         ]);
 
-        return redirect('/dosen/suratizin')->with("success","Data Berhasil Ditambah !");
+        return redirect('/dosen/suratizin/'.$request->id_prodi)->with("success","Data Berhasil Ditambah !");
     }
 
     public function edit($id){
@@ -145,14 +167,13 @@ class SuratIzinController extends Controller
             'judul_penelitian' => 'required|string|max:255',
             'dosen_pembimbing_1' => 'required|exists:dosen,nama',
             'dosen_pembimbing_2' => 'required|exists:dosen,nama',
-            'isi_surat' => 'required|string',
+            'isi_surat' => 'nullable|string',
         ],[
             'jenis.required' => 'Jenis wajib diisi.',
             'id_mahasiswa.exists' => 'Mahasiswa yang dipilih tidak valid..',
             'judul_penelitian.required' => 'Judul Penelitian wajib diisi.',
             'dosen_pembimbing_1.exists' => 'Dosen Bimbingan 1 yang dipilih tidak valid..',
             'dosen_pembimbing_2.exists' => 'Dosen Bimbingan 2 yang dipilih tidak valid..',
-            'isi_surat.required' => 'Isi surat wajib diisi.',
         ]);
         DB::table('surat_izin')  
             ->where('id', $id)
@@ -165,13 +186,21 @@ class SuratIzinController extends Controller
             'isi_surat' => $request->isi_surat
         ]);
 
-        return redirect('/dosen/suratizin')->with("success","Data Berhasil Diupdate !");
+        return redirect('/dosen/suratizin/'.$request->id_prodi)->with("success","Data Berhasil Diupdate !");
     }
 
     public function delete($id)
     {
-        DB::table('surat_izin')->where('id',$id)->delete();
+        // Ambil data berdasarkan ID
+        $suratizin = DB::table('surat_izin')->where('id', $id)->first();
 
-        return redirect('/dosen/suratizin')->with("success","Data Berhasil Dihapus !");
+        if (!$suratizin) {
+            return redirect()->back()->with('error', 'Data Surat Izin Penelitian tidak ditemukan!');
+        }
+
+        // Hapus jadwal
+        DB::table('surat_izin')->where('id', $id)->delete();
+
+        return redirect('/dosen/suratizin/'.$suratizin->id_prodi)->with("success","Data Berhasil Dihapus !");
     }
 }
